App/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
boto3==1.34.11
botocore==1.34.11
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
# command-not-found==0.3
cryptography==3.4.8
Cython==3.0.11
dbus-python==1.2.18
distro==1.7.0
distro-info===1.0
exceptiongroup==1.2.0
fastapi==0.105.0
Flask==3.0.0
Flask-Mail==0.10.0
gunicorn==23.0.0
# gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pillow==10.2.0
pydantic==2.5.2
pydantic_core==2.14.5
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pypng==0.20220715.0
# python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
s3transfer==0.10.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
systemd-python==234
translate==3.6.1
typing_extensions==4.9.0
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
urllib3==1.26.18
uvicorn==0.24.0.post1
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
